move reminder of dns funcs to util
